+++ src/main/java/kitchenpos/order/eatinorder/domain/OrderLineItem.java
@@ -1,8 +1,17 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.eatinorder.domain;
 
 import kitchenpos.menu.domain.Menu;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.ForeignKey;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.persistence.Transient;
 import java.math.BigDecimal;
 import java.util.UUID;
