+++ src/main/java/kitchenpos/order/eatinorder/domain/OrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.eatinorder.domain;
 
 import java.util.List;
 import java.util.Optional;
