+++ src/main/java/kitchenpos/order/eatinorder/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.eatinorder.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
