+++ src/main/java/kitchenpos/order/eatinorder/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.eatinorder.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
