+++ src/main/java/kitchenpos/order/eatinorder/ui/OrderRestController.java
@@ -1,9 +1,15 @@
-package kitchenpos.order.ui;
+package kitchenpos.order.eatinorder.ui;
 
-import kitchenpos.order.application.OrderService;
-import kitchenpos.order.domain.Order;
+import kitchenpos.order.eatinorder.application.OrderService;
+import kitchenpos.order.eatinorder.domain.Order;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
 import java.util.List;
